Move attachments and files models
