+++ botx/client/notifications_api/direct_notification.py
@@ -1,8 +1,6 @@
 from typing import Any, Dict, List, Optional, Union
 from uuid import UUID
 
-from botx.bot.models.outgoing_attachment import OutgoingAttachment
-from botx.client.attachments import BotXAPIAttachment
 from botx.client.authorized_botx_method import AuthorizedBotXMethod
 from botx.client.botx_method import callback_exception_thrower
 from botx.client.exceptions.common import ChatNotFoundError
@@ -22,11 +20,11 @@ from botx.client.notifications_api.mentions import (
     find_and_replace_embed_mentions,
 )
 from botx.missing import Missing, Undefined
+from botx.models.attachments import BotXAPIAttachment, IncomingFileAttachment, OutgoingAttachment
 from botx.shared_models.api_base import (
     UnverifiedPayloadBaseModel,
     VerifiedPayloadBaseModel,
 )
-from botx.shared_models.domain.attachments import IncomingFileAttachment
 
 try:
     from typing import Literal
